Add controller

+++ Entities/AppUser.cs
@@ -2,7 +2,7 @@
 {
     public class AppUser : BaseEntity
     {
-        public int ChatId { get; set; }
+        public long ChatId { get; set; }
         public string Username { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
